backend/tts/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-adk==1.27.4
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
boto3>=1.34.0
